refactor(frontend): extract layout shell and dashboard into separate files
App.tsx now only contains header/footer layout, AuthCallback and Dashboard are standalone pages. Removes duplicated UI code and cleans up footer.
