+++ frontend/src/pages/AuthCallback.tsx
@@ -1,6 +1,12 @@
 import { Button } from "@/components/ui/button";
-import { Card, CardContent } from "@/components/ui/card";
-import { CheckCircle, Loader2, XCircle } from "lucide-react";
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardHeader,
+  CardTitle,
+} from "@/components/ui/card";
+import { ArrowLeft, CheckCircle, Loader2, XCircle } from "lucide-react";
 import React, { useEffect } from "react";
 
 const AuthCallback: React.FC = () => {
@@ -13,16 +19,11 @@ const AuthCallback: React.FC = () => {
         const error = urlParams.get("error");
 
         if (success === "true" && encodedData) {
-          // Décoder les données
           const jsonData = atob(encodedData);
           const authData = JSON.parse(jsonData);
 
-          console.log("Auth data received:", authData);
-
-          // Stocker les données dans localStorage temporairement
           localStorage.setItem("google_auth_data", JSON.stringify(authData));
 
-          // Envoyer un message à la fenêtre parent si c'est une popup
           if (window.opener) {
             window.opener.postMessage(
               {
@@ -33,7 +34,6 @@ const AuthCallback: React.FC = () => {
             );
             window.close();
           } else {
-            // Rediriger vers la page principale après un délai
             setTimeout(() => {
               window.location.href = "/";
             }, 2000);
@@ -81,60 +81,60 @@ const AuthCallback: React.FC = () => {
   const error = urlParams.get("error");
 
   return (
-    <div className="min-h-screen flex items-center justify-center bg-background">
+    <div className="flex items-center justify-center flex-1">
       <Card className="w-full max-w-md">
-        <CardContent className="pt-6">
+        <CardHeader className="text-center">
           {success === "true" ? (
-            <div className="space-y-6 text-center">
-              <CheckCircle className="h-16 w-16 text-green-500 mx-auto" />
-              <div className="space-y-2">
-                <h2 className="text-2xl font-bold text-green-800">
-                  Authentification réussie !
-                </h2>
-                <p className="text-muted-foreground">
-                  Votre compte Google a été connecté avec succès.
-                </p>
+            <>
+              <div className="mx-auto mb-2">
+                <CheckCircle className="h-12 w-12 text-green-500" />
               </div>
-              <div className="flex items-center justify-center space-x-2 text-sm text-muted-foreground">
-                <Loader2 className="h-4 w-4 animate-spin" />
-                <span>Redirection automatique...</span>
-              </div>
-              <Button
-                variant="outline"
-                onClick={() => (window.location.href = "/")}
-                className="w-full"
-              >
-                Retourner à l'application
-              </Button>
-            </div>
+              <CardTitle>Authentification réussie</CardTitle>
+              <CardDescription>
+                Votre compte Google a été connecté avec succès.
+              </CardDescription>
+            </>
           ) : error ? (
-            <div className="space-y-6 text-center">
-              <XCircle className="h-16 w-16 text-destructive mx-auto" />
-              <div className="space-y-2">
-                <h2 className="text-2xl font-bold text-destructive">
-                  Erreur d'authentification
-                </h2>
-                <p className="text-muted-foreground">{error}</p>
+            <>
+              <div className="mx-auto mb-2">
+                <XCircle className="h-12 w-12 text-destructive" />
               </div>
-              <Button
-                variant="outline"
-                onClick={() => (window.location.href = "/")}
-                className="w-full"
-              >
-                Retourner à l'application
-              </Button>
-            </div>
+              <CardTitle className="text-destructive">
+                Erreur d'authentification
+              </CardTitle>
+              <CardDescription>{error}</CardDescription>
+            </>
           ) : (
-            <div className="space-y-6 text-center">
-              <Loader2 className="h-16 w-16 animate-spin text-primary mx-auto" />
-              <div className="space-y-2">
-                <h2 className="text-2xl font-bold">Traitement en cours...</h2>
-                <p className="text-muted-foreground">
-                  Veuillez patienter pendant que nous traitons votre
-                  authentification.
-                </p>
+            <>
+              <div className="mx-auto mb-2">
+                <Loader2 className="h-12 w-12 animate-spin text-primary" />
               </div>
+              <CardTitle>Traitement en cours...</CardTitle>
+              <CardDescription>
+                Veuillez patienter pendant que nous traitons votre
+                authentification.
+              </CardDescription>
+            </>
+          )}
+        </CardHeader>
+
+        <CardContent className="space-y-4">
+          {success === "true" && (
+            <div className="flex items-center justify-center gap-2 text-sm text-muted-foreground">
+              <Loader2 className="h-4 w-4 animate-spin" />
+              <span>Redirection automatique...</span>
             </div>
+          )}
+
+          {(success === "true" || error) && (
+            <Button
+              variant="outline"
+              onClick={() => (window.location.href = "/")}
+              className="w-full"
+            >
+              <ArrowLeft className="h-4 w-4" />
+              Retourner à l'application
+            </Button>
           )}
         </CardContent>
       </Card>
